modularized payment model by splitting original 1 files into 4 separate files for better structure and maintainability
