+++ backend/routes/index.js
@@ -34,7 +34,7 @@ const AppointmentRoutes = require('../modules/appointment/appointment.route');
 const PrescriptionRoutes = require('../modules/prescription/prescription.route');
 const MedicalRecordRoutes = require('../modules/medicalRecord/medicalRecord.route');
 const LabTestRoutes = require('../modules/labTest/labTest.route');
-// const PaymentRoutes = require('../modules/payment/payment.route');
+const PaymentRoutes = require('../modules/payment/payment.route');
 const AdminRoutes = require('../modules/admin/admin.route');
 const ChatRoutes = require('../modules/chat/chat.route');
 const AnalyticsRoutes = require('../modules/analytics/analytics.route');
@@ -50,10 +50,10 @@ router.use('/appointments', AppointmentRoutes);
 router.use('/prescriptions', PrescriptionRoutes);
 router.use('/medical-records', MedicalRecordRoutes);
 router.use('/lab-tests', LabTestRoutes);
-// router.use('/payments', PaymentRoutes);
+router.use('/payments', PaymentRoutes);
 router.use('/admin', AdminRoutes);
 router.use('/chat', ChatRoutes);
-// router.use('/analytics', AnalyticsRoutes);
+router.use('/analytics', AnalyticsRoutes);
 router.use('/health', HealthDataRoutes);
 router.use('/inventory', InventoryRoutes);
 router.use('/logs', LogRoutes);
